Doors implemented!
Press 'E' to open doors.
Doors can be set to manual open/close or automatic open/close

There is a bug with some of the doors. Although I copy and pasted the
doors, I suspect the rotation of the room affects the door opening. As a
result, I advice to separate doors from the rooms to avoid conflict.