Reorg of rooms and general fixes
Re-organized rooms. Removed doors from respective prefabs.
Doors are now inside the "DoorSet_Container" and are SEPERATE from the
rooms themselves.

Doors are proven to not work when Y rotation value is NOT set to "0".
Not sure why but yea.

And when naming things please make sure to have the following
conventions:
Block/Tile_Int/Ext_RespectiveName